manifest placeholders and some more fixes
