+++ build-logic/convention/build.gradle.kts
@@ -17,21 +17,13 @@ gradlePlugin {
             id = "ly.david.android.application"
             implementationClass = "ly.david.convention.plugin.AndroidApplicationConventionPlugin"
         }
-        register("androidTest") {
-            id = "ly.david.android.test"
-            implementationClass = "ly.david.convention.plugin.AndroidTestConventionPlugin"
-        }
         register("composeMultiplatform") {
             id = "ly.david.musicsearch.compose.multiplatform"
             implementationClass = "ly.david.convention.plugin.ComposeMultiplatformConventionPlugin"
         }
-        register("kotlinJvm") {
-            id = "ly.david.kotlin"
-            implementationClass = "ly.david.convention.plugin.KotlinJvmConventionPlugin"
-        }
         register("kotlinMultiplatform") {
             id = "ly.david.musicsearch.kotlin.multiplatform"
-            implementationClass = "ly.david.convention.plugin.KotlinMultiplatformConventionPlugin"
+            implementationClass = "ly.david.convention.plugin.KotlinMultiplatformLibraryConventionPlugin"
         }
         register("parcelize") {
             id = "ly.david.parcelize"
