--- build-logic/convention/src/main/kotlin/ly/david/convention/ConfigureKotlin.kt
@@ -1,16 +1,0 @@
-package ly.david.convention
-
-import org.gradle.api.Project
-import org.gradle.api.plugins.JavaPluginExtension
-import org.gradle.jvm.toolchain.JavaLanguageVersion
-import org.gradle.kotlin.dsl.configure
-
-private const val JAVA_VERSION = 21
-
-fun Project.configureKotlin() {
-    extensions.configure<JavaPluginExtension> {
-        toolchain {
-            languageVersion.set(JavaLanguageVersion.of(JAVA_VERSION))
-        }
-    }
-}
