+++ build-logic/convention/src/main/kotlin/ly/david/convention/plugin/AndroidApplicationConventionPlugin.kt
@@ -1,9 +1,14 @@
 package ly.david.convention.plugin
 
-import ly.david.convention.configureAndroid
+import com.android.build.api.dsl.ApplicationExtension
 import ly.david.convention.configureDetekt
+import org.gradle.api.JavaVersion
 import org.gradle.api.Plugin
 import org.gradle.api.Project
+import org.gradle.kotlin.dsl.configure
+
+const val COMPILE_SDK_VERSION = 36
+private const val MIN_SDK_VERSION = 24
 
 @Suppress("unused")
 class AndroidApplicationConventionPlugin : Plugin<Project> {
@@ -19,3 +24,27 @@ class AndroidApplicationConventionPlugin : Plugin<Project> {
         }
     }
 }
+
+private fun Project.configureAndroid() {
+    android {
+        compileSdk = COMPILE_SDK_VERSION
+
+        defaultConfig {
+            minSdk = MIN_SDK_VERSION
+
+            manifestPlaceholders += mapOf("appAuthRedirectScheme" to "")
+        }
+
+        compileOptions {
+            sourceCompatibility = JavaVersion.VERSION_21
+            targetCompatibility = JavaVersion.VERSION_21
+//                isCoreLibraryDesugaringEnabled = true
+        }
+
+//            dependencies {
+//                add("coreLibraryDesugaring", libs.findLibrary("desugarjdklibs").get())
+//            }
+    }
+}
+
+private fun Project.android(configure: ApplicationExtension.() -> Unit) = extensions.configure(configure)
