+++ shared/feature/history/src/androidHostTest/snapshots/images/ly.david.musicsearch.shared.feature.history_HistoryListItemTest_artistWithCoverArt[NIGHT].png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:0700d911db451699eb9197a0209534662c57c419ce5972947bd1ea2c9acefe84
+size 15184
