+++ shared/feature/history/src/androidHostTest/snapshots/images/ly.david.musicsearch.shared.feature.history_HistoryListItemTest_artistWithCoverArt[NOTNIGHT].png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:be26d2e3bb8ed7396cf6b3b943da4035e476bf9a0b02b7eb4e96006cac9eb657
+size 14746
