+++ shared/feature/history/src/androidHostTest/snapshots/images/ly.david.musicsearch.shared.feature.history_HistoryListItemTest_artist[NIGHT].png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:838473f9f613bfebbe355cfdb3d87c9d4bd91b1231c20cde348c579b04f42c21
+size 13389
