+++ shared/feature/history/src/androidHostTest/snapshots/images/ly.david.musicsearch.shared.feature.history_HistoryListItemTest_artist[NOTNIGHT].png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:74235bfc40f3d983dc8f4d4422f5306eac7fe9fd0f17b249f511d77048f949a9
+size 13356
