+++ shared/feature/history/src/androidHostTest/snapshots/images/ly.david.musicsearch.shared.feature.history_HistoryListItemTest_releaseGroupWithCoverArt[NIGHT].png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:eaa3fffe4569c0d0a5279c27e284e41c025a35af431c39008733e8fbce06873c
+size 25020
