+++ shared/feature/history/src/androidHostTest/snapshots/images/ly.david.musicsearch.shared.feature.history_HistoryListItemTest_releaseGroupWithCoverArt[NOTNIGHT].png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:9f7cbf25f93ec830f75e061cbaa858cb56d2f7747524a736c657a101c5fa25ca
+size 24949
