+++ shared/feature/history/src/androidHostTest/snapshots/images/ly.david.musicsearch.shared.feature.history_HistoryListItemTest_releaseWithCoverArt[NIGHT].png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:243fe35da3d4a94e5dc4ded65c132f08704e6f46e069bf4d6a846b55d939df05
+size 20778
