+++ shared/feature/history/src/androidHostTest/snapshots/images/ly.david.musicsearch.shared.feature.history_HistoryListItemTest_releaseWithCoverArt[NOTNIGHT].png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:50c5fda70346d9754f3a078f6b2dafc24a216ef2f1047162f7dbfabd7ee0e4d8
+size 20778
