+++ shared/feature/history/src/commonMain/kotlin/ly/david/musicsearch/shared/feature/history/internal/HistoryPresenter.kt
@@ -112,7 +112,9 @@ internal class HistoryPresenter(
             sortOption = sortOption,
             pagingDataFlow = pagingDataFlow,
             lazyListState = lazyListState,
-            eventSink = ::eventSink,
+            eventSink = { event ->
+                eventSink(event)
+            },
         )
     }
 }
