--- shared/feature/history/src/test/snapshots/images/ly.david.musicsearch.shared.feature.history_HistoryListItemTest_artistWithCoverArt[NIGHT].png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:9facd72ffb635fe33f409838b371b492de4c04ef6163005fe219f24ffcc9956e
-size 15475
