--- shared/feature/history/src/test/snapshots/images/ly.david.musicsearch.shared.feature.history_HistoryListItemTest_artistWithCoverArt[NOTNIGHT].png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:12667842a504c2b248eecd2becf2bfc3420db0f137c09a844403b2dfa5e2d277
-size 15056
