--- shared/feature/history/src/test/snapshots/images/ly.david.musicsearch.shared.feature.history_HistoryListItemTest_artist[NIGHT].png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:8727a5664316d0dc764c933ac373982864e1eebdf095c482d6f14ae15785a540
-size 13691
