--- shared/feature/history/src/test/snapshots/images/ly.david.musicsearch.shared.feature.history_HistoryListItemTest_artist[NOTNIGHT].png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:368632243ca1da061c7e4ba10d54c7216c9fdd94103d50b327bc7e17b3041a82
-size 13705
