--- shared/feature/history/src/test/snapshots/images/ly.david.musicsearch.shared.feature.history_HistoryListItemTest_releaseGroupWithCoverArt[NIGHT].png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:3ed9d0311b47de0698d510eda117ac254a9cda617f5d66406221c29bd3359967
-size 25670
