--- shared/feature/history/src/test/snapshots/images/ly.david.musicsearch.shared.feature.history_HistoryListItemTest_releaseGroup[NIGHT].png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:949af5acbff1c0f256cff167cc6b471c77083c3b9803454f5c57a804b1bfd135
-size 26206
