--- shared/feature/history/src/test/snapshots/images/ly.david.musicsearch.shared.feature.history_HistoryListItemTest_releaseGroup[NOTNIGHT].png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:e740353bc29017b759a1852afe7a88f179a7b2bb7bacbeec0cea794e7cb42a8e
-size 26232
