--- shared/feature/history/src/test/snapshots/images/ly.david.musicsearch.shared.feature.history_HistoryListItemTest_releaseWithCoverArt[NIGHT].png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:4286c4ba32ed0283ab003768f686689fe501294113768bc747cd80aca7c38198
-size 21438
