--- shared/feature/history/src/test/snapshots/images/ly.david.musicsearch.shared.feature.history_HistoryListItemTest_releaseWithCoverArt[NOTNIGHT].png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:f62824831efca53ae1e030f967b6ca955827052183bad03cf008d1c5ceaf8ae4
-size 21367
