--- shared/feature/history/src/test/snapshots/images/ly.david.musicsearch.shared.feature.history_HistoryListItemTest_release[NIGHT].png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:8b9e70459d989aa1efc33c848baba2410232fa50454deff7d655f4be2bcc1a32
-size 25114
