--- shared/feature/history/src/test/snapshots/images/ly.david.musicsearch.shared.feature.history_HistoryListItemTest_release[NOTNIGHT].png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:84224d35bec442fa6d77a37c115599e3fead760fdfe542dba1b0a6022352ea02
-size 25059
