--- shared/feature/history/src/test/snapshots/images/ly.david.musicsearch.shared.feature.history_HistoryListItemTest_withFilter[NIGHT].png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:77794b742230b48c4ee2debf0ef993970aaf5b6990045879382de8838cff8fdc
-size 22487
