--- shared/feature/history/src/test/snapshots/images/ly.david.musicsearch.shared.feature.history_HistoryListItemTest_withFilter[NOTNIGHT].png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:ccd9d9044edf74cae26d8ea6999bb1123f4e3cf118f7a6c2b5266291c30912e4
-size 22401
